Improve test coverage ✅ - Refactor/Move ReferenceAttachmentResolverTest to engine module - Enable disabled DefaultRepository unit test - Fix failing unit test
